# src/protos/video/VideoUploadResponse.ts
// Original file: src/protos/video.proto


export interface VideoUploadResponse {
  'id'?: (string);
  'size'?: (number);
}

export interface VideoUploadResponse__Output {
  'id'?: (string);
  'size'?: (number);
}